improve logic & clippy warning

+++ pallets/staking/src/inflation.rs
@@ -41,7 +41,7 @@ where
 	// Milliseconds per year for the Julian year (365.25 days).
 	const MILLISECONDS_PER_YEAR: u64 = 1000 * 3600 * 24 * 36525 / 100;
 
-	let portion = Perbill::from_rational(era_duration as u64, MILLISECONDS_PER_YEAR);
+	let portion = Perbill::from_rational(era_duration, MILLISECONDS_PER_YEAR);
 	let payout = portion
 		* yearly_inflation
 			.calculate_for_fraction_times_denominator(npos_token_staked, total_tokens.clone());
